Split up responder
